Reestruturação do código para facilitar o desenvolvimento, manutenção e colaboração
